Stuck for now
Updated the button and text inputs to send values to a variable. Not sure how to display the results in the table. Debating whether or not to clear the inputs upon search, seems like extra, but I'd like to do it. Will pick up tomorrow learning how to deal with table inputs in React